requirements.txt
requests>=2.32.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
umap-learn>=0.5.6
plotly>=5.24.0
streamlit>=1.40.0
tqdm>=4.66.0
